check the number of tiles, before merging

+++ src/loader.py
@@ -313,11 +313,22 @@ def load_raster_file(entry: Entry, executor: Executor) -> str:
     tiles = [future.result() for future in futures if future.result() is not None]
     
     # run the merge function and delete the other files
-    _wbt_merge_raster(dataset_base_path, f"{entry.variable.name.replace(' ', '_')}_{entry.id}.tif")
+    if len(tiles) > 1:
+        logger.debug('Starting WhitboxTools mosaic operation...')
+        _wbt_merge_raster(dataset_base_path, f"{entry.variable.name.replace(' ', '_')}_{entry.id}.tif")
 
-    # remove the tiles
-    for tile in tiles:
-        Path(tile).unlink()
+        # remove the tiles
+        for tile in tiles:
+            Path(tile).unlink()
+    
+    # check if there is exactly one tile
+    elif len(tiles) == 1:
+        # rename the file
+        new_name = dataset_base_path / f"{entry.variable.name.replace(' ', '_')}_{entry.id}.tif"
+        Path(tiles[0]).rename(new_name)
+        tiles = [str(new_name)]
+    else:
+        logger.warning(f'No tiles were clipped for the reference area. It might not be covered by dataset <ID={entry.id}>')
 
     # save the metadata
     metafile_name = str(params.dataset_path / f"{entry.variable.name.replace(' ', '_')}_{entry.id}.metadata.json")
@@ -339,7 +350,7 @@ def _rio_clip_raster(file_name: str, reference_area: gpd.GeoDataFrame, base_path
         except ValueError as e:
             if 'Input shapes do not overlap raster' in str(e):
                 logger.debug(f"Skipping {file_name} as it does not overlap with the reference area.")
-                return
+                return None
             else:
                 logger.exception(f"An unexpected error occured: {str(e)}")
